Refactor symmetric operations.
