+++ lib/constants.js
@@ -3,7 +3,10 @@
  */
 'use strict';
 
-exports.keyTypes = {
+exports.AES_KW_ALGORITHM = 'id-aes256-wrap';
+exports.AES_KW_RFC3394_IV = Buffer.from('A6A6A6A6A6A6A6A6', 'hex');
+
+exports.KEY_TYPES = {
   asymmetric: new Set([
     'Ed25519VerificationKey2018',
   ]),
