+++ lib/index.js
@@ -7,19 +7,14 @@ const _asymmetricKey = require('./asymmetric-key');
 const _constants = require('./constants');
 const _symmetricKey = require('./symmetric-key');
 const assert = require('assert-plus');
-const base64url = require('base64url-universal');
 const bedrock = require('bedrock');
 const brPackageManager = require('bedrock-package-manager');
-const crypto = require('crypto');
 const database = require('bedrock-mongodb');
 const {promisify} = require('util');
 const {util: {BedrockError}} = bedrock;
 
 // load config defaults
 require('./config');
-
-const AES_KW_ALGORITHM = 'id-aes256-wrap';
-const AES_KW_RFC3394_IV = Buffer.from('A6A6A6A6A6A6A6A6', 'hex');
 
 bedrock.events.on('bedrock.init', () => brPackageManager.register({
   alias: 'ssm-v1',
@@ -58,9 +53,9 @@ exports.generateKey = async ({keyId, operation}) => {
   const {invocationTarget: {type}} = operation;
   assert.string(type, 'options.operation.type');
 
-  if(_constants.keyTypes.asymmetric.has(type)) {
+  if(_constants.KEY_TYPES.asymmetric.has(type)) {
     return _asymmetricKey.generateKeyPair({keyId, type});
-  } else if(_constants.keyTypes.symmetric.has(type)) {
+  } else if(_constants.KEY_TYPES.symmetric.has(type)) {
     return _symmetricKey.generateKey({keyId, type});
   }
 
@@ -81,13 +76,7 @@ exports.wrapKey = async ({keyId, operation}) => {
   assert.object(operation, 'options.operation');
 
   const {key: kek} = await _getKeyRecord({id: keyId});
-  if(kek.type !== 'AesKeyWrappingKey2019') {
-    throw new Error(`Unknown key type "${kek.type}".`);
-  }
-
-  const {unwrappedKey} = operation;
-  const wrappedKey = await _aesWrapKey({kek, unwrappedKey});
-  return {wrappedKey};
+  return _symmetricKey.wrapKey({kek, operation});
 };
 
 /**
@@ -104,13 +93,7 @@ exports.unwrapKey = async ({keyId, operation}) => {
   assert.object(operation, 'options.operation');
 
   const {key: kek} = await _getKeyRecord({id: keyId});
-  if(kek.type !== 'AesKeyWrappingKey2019') {
-    throw new Error(`Unknown key type "${kek.type}".`);
-  }
-
-  const {wrappedKey} = operation;
-  const unwrappedKey = await _aesUnwrapKey({kek, wrappedKey});
-  return {unwrappedKey};
+  return _symmetricKey.unwrapKey({kek, operation});
 };
 
 /**
@@ -132,9 +115,9 @@ exports.sign = async ({keyId, operation}) => {
   const {key} = await _getKeyRecord({id: keyId});
   const {type} = key;
 
-  if(_constants.keyTypes.asymmetric.has(type)) {
+  if(_constants.KEY_TYPES.asymmetric.has(type)) {
     return _asymmetricKey.sign({key, operation});
-  } else if(_constants.keyTypes.symmetric.has(type)) {
+  } else if(_constants.KEY_TYPES.symmetric.has(type)) {
     return _symmetricKey.sign({key, operation});
   }
   // this should never happen
@@ -161,9 +144,9 @@ exports.verify = async ({keyId, operation}) => {
   const {key} = await _getKeyRecord({id: keyId});
   const {type} = key;
 
-  if(_constants.keyTypes.symmetric.has(type)) {
+  if(_constants.KEY_TYPES.symmetric.has(type)) {
     return _symmetricKey.verify({key, operation});
-  } else if(_constants.keyTypes.asymmetric.has(type)) {
+  } else if(_constants.KEY_TYPES.asymmetric.has(type)) {
     throw new Error(`Unsupported operation for key type "${type}".`);
   }
   // this should never happen
@@ -193,22 +176,4 @@ async function _getKeyRecord({id}) {
   }
 
   return record;
-}
-
-async function _aesWrapKey({kek, unwrappedKey}) {
-  unwrappedKey = base64url.decode(unwrappedKey);
-  const secret = base64url.decode(kek.secret);
-  const cipher = crypto.createCipheriv(
-    AES_KW_ALGORITHM, secret, AES_KW_RFC3394_IV);
-  const output = Buffer.concat([cipher.update(unwrappedKey), cipher.final()]);
-  return base64url.encode(output);
-}
-
-async function _aesUnwrapKey({kek, wrappedKey}) {
-  wrappedKey = base64url.decode(wrappedKey);
-  const secret = base64url.decode(kek.secret);
-  const decipher = crypto.createDecipheriv(
-    AES_KW_ALGORITHM, secret, AES_KW_RFC3394_IV);
-  const output = Buffer.concat([decipher.update(wrappedKey), decipher.final()]);
-  return base64url.encode(output);
 }
