+++ lib/symmetric-key.js
@@ -3,6 +3,7 @@
  */
 'use strict';
 
+const _constants = require('./constants');
 const base64url = require('base64url-universal');
 const crypto = require('crypto');
 const database = require('bedrock-mongodb');
@@ -99,6 +100,62 @@ exports.verify = async ({key, operation}) => {
   const verified = await _hs256Verify({key, verifyData, signatureValue});
   return {verified};
 };
+
+/**
+ * Wraps a cryptographic key using a key encryption key (KEK).
+ *
+ * @ignore
+ * @param {Object} options - The options to use.
+ * @param {string} options.kek - The key encryption key to use.
+ * @param {Object} options.operation - The KMS operation.
+ *
+ * @returns {Promise<Object>} An object containing `{wrappedKey}`.
+ */
+exports.wrapKey = async ({kek, operation}) => {
+  if(kek.type !== 'AesKeyWrappingKey2019') {
+    throw new Error(`Unknown key type "${kek.type}".`);
+  }
+  const {unwrappedKey} = operation;
+  const wrappedKey = await _aesWrapKey({kek, unwrappedKey});
+  return {wrappedKey};
+};
+
+/**
+ * Unwraps a cryptographic key using a key encryption key (KEK).
+ *
+ * @ignore
+ * @param {Object} options - The options to use.
+ * @param {string} options.kek - The key encryption key to use.
+ * @param {Object} options.operation - The KMS operation.
+ *
+ * @returns {Promise<Object>} An object containing `{unwrappedKey}`.
+ */
+exports.unwrapKey = async ({kek, operation}) => {
+  if(kek.type !== 'AesKeyWrappingKey2019') {
+    throw new Error(`Unknown key type "${kek.type}".`);
+  }
+  const {wrappedKey} = operation;
+  const unwrappedKey = await _aesUnwrapKey({kek, wrappedKey});
+  return {unwrappedKey};
+};
+
+async function _aesWrapKey({kek, unwrappedKey}) {
+  unwrappedKey = base64url.decode(unwrappedKey);
+  const secret = base64url.decode(kek.secret);
+  const cipher = crypto.createCipheriv(
+    _constants.AES_KW_ALGORITHM, secret, _constants.AES_KW_RFC3394_IV);
+  const output = Buffer.concat([cipher.update(unwrappedKey), cipher.final()]);
+  return base64url.encode(output);
+}
+
+async function _aesUnwrapKey({kek, wrappedKey}) {
+  wrappedKey = base64url.decode(wrappedKey);
+  const secret = base64url.decode(kek.secret);
+  const decipher = crypto.createDecipheriv(
+    _constants.AES_KW_ALGORITHM, secret, _constants.AES_KW_RFC3394_IV);
+  const output = Buffer.concat([decipher.update(wrappedKey), decipher.final()]);
+  return base64url.encode(output);
+}
 
 async function _hs256Sign({key, verifyData}) {
   const secret = base64url.decode(key.secret);
